requirements.txt
uwsgi==2.0.18
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-SocketIO==4.3.0
Flask-SSLify==0.1.5
hurry.filesize==0.9
eventlet==0.25.2
gevent==20.6.1
gevent-websocket==0.10.1
greenlet==0.4.16